add Enum
